+++ v1/services/ecosistemas.services.js
@@ -1,10 +1,16 @@
 import Ecosistema from "../models/ecosistema.model.js";
+import { isValidObjectId } from "mongoose";
+
+
 
 export const obtenerEcosistemasService = async () => {
   return await Ecosistema.find();
 };
 
 export const obtenerEcosistemaPorIdService = async (id) => {
+  if (!isValidObjectId(id)) {
+    throw new Error("ID inválido");
+  }
   return await Ecosistema.findById(id);
 };
 
@@ -15,10 +21,16 @@ export const crearEcosistemaService = async (ecosistemaGuardar) => {
 };
 
 export const actualizarEcosistemaService = async (id, ecosistemaActualizar) => {
+  if (!isValidObjectId(id)) {
+    throw new Error("ID inválido");
+  }
   return await Ecosistema.findByIdAndUpdate(id, ecosistemaActualizar, { new: true });
 };
 
 export const eliminarEcosistemaService = async (id) => {
+  if (!isValidObjectId(id)) {
+    throw new Error("ID inválido");
+  }
   return await Ecosistema.findByIdAndDelete(id);
 };
